Create SelectWorkspace component and use it in result views
