+++ src/codeclarity_components/results/sbom/SbomContent.vue
@@ -18,16 +18,7 @@ import { useAuthStore } from '@/stores/auth';
 import { ResultsRepository } from '@/codeclarity_components/results/results.repository';
 import type { DataResponse } from '@/utils/api/responses/DataResponse';
 import SbomTable from './SbomTable.vue';
-import { WorkspacesOutput } from '../workspace.entity';
-import {
-    Select,
-    SelectContent,
-    SelectGroup,
-    SelectItem,
-    SelectLabel,
-    SelectTrigger,
-    SelectValue
-} from '@/shadcn/ui/select';
+import SelectWorkspace from '../SelectWorkspace.vue';
 
 export interface Props {
     analysisID?: string;
@@ -52,7 +43,6 @@ const loading: Ref<boolean> = ref(true);
 
 const render: Ref<boolean> = ref(false);
 const stats: Ref<SbomStats> = ref(new SbomStats());
-const workspaces: Ref<WorkspacesOutput> = ref(new WorkspacesOutput());
 const selected_workspace: Ref<string> = ref('.');
 
 watch(
@@ -128,28 +118,6 @@ async function getSbomStats(refresh: boolean = false) {
         createDepStatusDistChart();
     }
 }
-
-async function getSbomWorkspaces() {
-    if (!userStore.getDefaultOrg) return;
-    if (!(authStore.getAuthenticated && authStore.getToken)) return;
-    try {
-        const res = await sbomRepo.getSbomWorkspaces({
-            orgId: userStore.getDefaultOrg.id,
-            projectId: props.projectID,
-            analysisId: props.analysisID,
-            bearerToken: authStore.getToken,
-            handleBusinessErrors: true
-        });
-        workspaces.value = res.data;
-    } catch (_err) {
-        console.error(_err);
-        error.value = true;
-        // if (_err instanceof BusinessLogicError) {
-        //     errorCode.value = _err.error_code;
-        // }
-    }
-}
-getSbomWorkspaces();
 
 // Create charts
 function createDepStatusDistChart() {
@@ -249,32 +217,15 @@ function createDepTypeChart() {
 
 <template>
     <div value="sbom" class="space-y-4">
-        <div class="flex flex-col items-center">
-            <Select
-                @update:model-value="
-                    (e) => {
-                        selected_workspace = e as string;
-                        getSbomStats();
-                    }
-                "
-            >
-                <SelectTrigger class="w-[180px]">
-                    <SelectValue placeholder="Select a workspace" />
-                </SelectTrigger>
-                <SelectContent>
-                    <SelectGroup>
-                        <SelectLabel>Workspaces</SelectLabel>
-                        <SelectItem
-                            v-for="workspace of workspaces.workspaces"
-                            :key="workspace"
-                            :value="workspace"
-                        >
-                            {{ workspace }}
-                        </SelectItem>
-                    </SelectGroup>
-                </SelectContent>
-            </Select>
-        </div>
+        
+        <SelectWorkspace
+            :projectID="projectID"
+            :analysisID="analysisID"
+            :get-function="getSbomStats"
+            v-model:error="error"
+            v-model:selected_workspace="selected_workspace"
+        ></SelectWorkspace>
+
         <div class="grid gap-4 sm:grid-cols-2 lg:grid-cols-8">
             <Card class="lg:col-start-3">
                 <CardHeader class="flex flex-col items-center">
